+++ NEWS.md
@@ -1,3 +1,10 @@
+# chipmunkbasic 0.1.1
+
+* Feature: Add convex polygons
+* Feature: Can now set initial angular velocity
+* Vignette: Added "Plop Polygon" vignette
+
 # chipmunkbasic 0.1.0
 
 * Initial release
+* Split the R6 wrapper from `chipmunkcore` into this package